feat: fix hidden AAAmit pr comments
