--- frontend/src/hooks/loginCredentials.tsx
@@ -1,7 +1,0 @@
-import { useState } from "react";
-
-export function useLoginCredentials() {
-  const [data, setData] = useState({ username: "", password: "" });
-
-  return {data, setData};
-}
